Unify controller mapping index types

+++ src/main/PAL/main/pad/begin_controller_config.c
@@ -11,6 +11,6 @@ void BeginControllerConfig(void) {
     g_ControllerSceneAngleX = 0;
     g_PadConfigFlipTimer = 0;
     g_PadConfigFlipPhase = 0;
-    g_PadMappingIndexSaved = g_PadMappingIndex;
-    g_NegconMappingIndexSaved = g_NegconMappingIndex;
+    g_PadMappingIndexSaved = (u16)g_PadMappingIndex;
+    g_NegconMappingIndexSaved = (u16)g_NegconMappingIndex;
 }
